+++ R/shiny_priors.R
@@ -10,17 +10,32 @@
 #'
 #' @return shiny::app
 #' @keywords internal
-priors_shiny <- function(spatial_data,
-                         measurement_data,
-                         time_variable,
-                         mesh,
-                         log_folder = NULL) {
+model_builder_shiny <- function(spatial_data,
+                                measurement_data,
+                                time_variable,
+                                mesh,
+                                log_folder = NULL) {
     future::plan(future::multisession())
 
     got_coords <- has_coords(spatial_data = spatial_data)
     if (!got_coords) {
         stop("Please make sure you have set coordinates on spatial_data using sp::coordinates.")
     }
+
+    spatial_crs <- sp::proj4string(spatial_data)
+    mesh_crs <- mesh$crs$input
+
+    if (is.na(mesh_crs) && is.na(spatial_crs)) {
+        warning("Cannot read CRS from mesh or spatial_data, using default CRS = +proj=longlat +datum=WGS84")
+        crs <- "+proj=longlat +datum=WGS84"
+    } else if (is.na(mesh_crs)) {
+        crs <- spatial_crs
+    } else {
+        crs <- mesh_crs
+    }
+
+    brewer_palettes <- RColorBrewer::brewer.pal.info
+    default_colours <- rownames(brewer_palettes[brewer_palettes$cat == "seq", ])
 
     # Text for priors help
     prior_range_text <- "A length 2 vector, with (range0, Prange) specifying that P(ρ < ρ_0)=p_ρ,
@@ -128,8 +143,17 @@ priors_shiny <- function(spatial_data,
                     type = "tabs",
                     shiny::tabPanel(
                         "Features",
-                        shiny::selectInput(inputId = "model_var", label = "Model variable", choices = features),
-                        shiny::selectInput(inputId = "exposure_param", label = "Exposure param", choices = features),
+                        shiny::fluidRow(
+                            shiny::column(
+                                6,
+                                shiny::selectInput(inputId = "model_var", label = "Model variable", choices = features),
+                                shiny::selectInput(inputId = "exposure_param", label = "Exposure (time variable)", choices = features),
+                            ),
+                            shiny::column(
+                                6,
+                                shiny::selectInput(inputId = "data_dist", label = "Data distribution", choices = c("Poisson", "Gaussian")),
+                            )
+                        ),
                         shiny::checkboxGroupInput(inputId = "features", label = "Features", choices = features),
                         shiny::checkboxInput(inputId = "f_func", label = "Add f()", value = FALSE),
                         shiny::actionButton(inputId = "clear", label = "Clear"),
@@ -158,8 +182,33 @@ priors_shiny <- function(spatial_data,
                         shiny::plotOutput(outputId = "plot_model_out")
                     ),
                     shiny::tabPanel(
+                        "Map",
+                        shiny::fluidRow(
+                            shiny::column(
+                                6,
+                                shiny::selectInput(inputId = "map_plot_type", label = "Plot type", choices = c("Predicted mean fields", "Random effect fields"), selected = "Predicted mean fields"),
+                                shiny::selectInput(inputId = "select_run_map", label = "Select run:", choices = c())
+                            ),
+                            shiny::column(
+                                6,
+                                shiny::selectInput(
+                                    inputId = "colour_category",
+                                    label = "Palette type",
+                                    choices = c("Sequential", "Diverging", "Qualitative", "Viridis"),
+                                    selected = "Viridis"
+                                ),
+                                shiny::selectInput(
+                                    inputId = "colour_scheme",
+                                    label = "Colour Scheme",
+                                    choices = default_colours,
+                                ),
+                            )
+                        ),
+                        leaflet::leafletOutput(outputId = "map_out")
+                    ),
+                    shiny::tabPanel(
                         "Code",
-                        shiny::selectInput(inputId = "select_run", label = "Select run:", choices = c()),
+                        shiny::selectInput(inputId = "select_run_code", label = "Select run:", choices = c()),
                         shiny::verbatimTextOutput(outputId = "code_out")
                     ),
                     shiny::tabPanel(
@@ -201,11 +250,9 @@ priors_shiny <- function(spatial_data,
         })
 
         shiny::observe({
-            shiny::updateSelectInput(session = session, inputId = "select_run", choices = run_names())
-        })
-
-        shiny::observeEvent(input$features, {
-            print(paste0("You have chosen: ", input$features))
+            shiny::updateSelectInput(session = session, inputId = "select_run_map", choices = run_names())
+            shiny::updateSelectInput(session = session, inputId = "select_run_code", choices = run_names())
+            shiny::updateSelectInput(session, inputId = "colour_scheme", label = "Colours", choices = category_colours())
         })
 
         shiny::observeEvent(input$clear, {
@@ -273,10 +320,20 @@ priors_shiny <- function(spatial_data,
             formula_str()
         })
 
+        data_distribution <- shiny::reactive({
+            tolower(input$data_dist)
+        })
+
         shiny::observeEvent(input$run_model, ignoreNULL = TRUE, {
             exposure_param_local <- input$exposure_param
             formula_local <- inla_formula()
             measurement_data_local <- measurement_data
+
+            data_dist_local <- data_distribution()
+            family_control <- NULL
+            if (data_dist_local == "poisson") {
+                family_control <- list(link = "log")
+            }
 
             promise <- promises::future_promise(
                 {
@@ -284,9 +341,9 @@ priors_shiny <- function(spatial_data,
                     require("INLA")
                     inlabru::bru(formula_local,
                         data = measurement_data_local,
-                        family = "poisson",
+                        family = data_dist_local,
                         E = measurement_data_local[[exposure_param_local]],
-                        control.family = list(link = "log"),
+                        control.family = family_control,
                         options = list(
                             verbose = FALSE
                         )
@@ -300,8 +357,10 @@ priors_shiny <- function(spatial_data,
                     function(model_output) {
                         # Run the model
                         run_no(run_no() + 1)
-                        model_vals$model_outputs[[run_no()]] <- model_output
-                        model_vals$parsed_outputs[[run_no()]] <- parse_model_output(
+                        run_label <- paste0("Run-", run_no())
+
+                        model_vals$model_outputs[[run_label]] <- model_output
+                        model_vals$parsed_outputs[[run_label]] <- parse_model_output(
                             model_output = model_output,
                             measurement_data = measurement_data
                         )
@@ -316,7 +375,6 @@ priors_shiny <- function(spatial_data,
                             "pg_ar1" = input$pg_ar1
                         )
 
-                        run_label <- paste0("Run-", run_no())
                         model_vals$run_params[[run_label]] <- run_params
 
                         if (write_logs) {
@@ -354,6 +412,69 @@ priors_shiny <- function(spatial_data,
             },
             rownames = TRUE
         )
+
+        category_colours <- shiny::reactive({
+            if (input$colour_category == "Viridis") {
+                colours <- c("viridis", "magma", "inferno", "plasma")
+            } else {
+                palettes_mapping <- list("Sequential" = "seq", "Diverging" = "div", "Qualitative" = "qual")
+                chosen_cat <- palettes_mapping[input$colour_category]
+                colours <- rownames(subset(RColorBrewer::brewer.pal.info, category %in% chosen_cat))
+            }
+            colours
+        })
+
+        colour_scheme <- shiny::reactive({
+            input$colour_scheme
+        })
+
+
+        prediction_field <- shiny::reactive({
+            if (length(model_vals$parsed_outputs) == 0) {
+                return()
+            }
+
+            data <- model_vals$parsed_outputs[[input$select_run_map]]
+            if (input$map_plot_type == "Predicted mean fields") {
+                create_prediction_field(
+                    mesh = mesh,
+                    plot_type = "predicted_mean_fields",
+                    data_dist = data_distribution(),
+                    var_a = data[["mean_post"]],
+                    var_b = data[["fixed_mean"]]
+                )
+            } else {
+                create_prediction_field(
+                    mesh = mesh,
+                    plot_type = "random_effect_fields",
+                    data_dist = data_distribution(),
+                    var_a = data[["mean_post"]]
+                )
+            }
+        })
+
+        z_values <- shiny::reactive({
+            prediction_field()[["z"]]
+        })
+
+        map_raster <- shiny::reactive({
+            raster::rasterFromXYZ(prediction_field(), crs = crs)
+        })
+
+        map_colours <- shiny::reactive({
+            leaflet::colorNumeric(palette = colour_scheme(), domain = z_values(), reverse = FALSE)
+        })
+
+        output$map_out <- leaflet::renderLeaflet({
+            if (is.null(map_raster())) {
+                return()
+            }
+
+            leaflet::leaflet() %>%
+                leaflet::addTiles(group = "OSM") %>%
+                leaflet::addRasterImage(map_raster(), colors = map_colours(), opacity = 0.9, group = "Raster") %>%
+                leaflet::addLegend(position = "topright", pal = map_colours(), values = z_values())
+        })
 
         model_plot <- shiny::eventReactive(input$plot_type, ignoreNULL = FALSE, {
             if (length(model_vals$parsed_outputs) == 0) {
@@ -401,7 +522,12 @@ priors_shiny <- function(spatial_data,
                 return()
             }
 
-            params <- model_vals$run_params[[input$select_run]]
+            params <- model_vals$run_params[[input$select_run_code]]
+
+            family_control_str <- "NULL"
+            if (data_distribution() == "poisson") {
+                family_control_str <- "list(link = 'log'),"
+            }
 
             paste0(
                 "spde <- INLA::inla.spde2.pcmatern(
@@ -416,9 +542,9 @@ priors_shiny <- function(spatial_data,
             )", "\n\n",
                     paste0("model_output <- inlabru::bru(formula,
                         data = measurement_data,
-                        family = 'poisson',
+                        family = '", data_distribution(), "',
                         E = measurement_data[[", input$exposure_param, "]],
-                        control.family = list(link = 'log'),
+                        control.family = ", family_control_str, "
                         options = list(
                             verbose = FALSE
                         )
@@ -441,6 +567,6 @@ priors_shiny <- function(spatial_data,
 #'
 #' @return shiny::app
 #' @export
-interactive_priors <- function(spatial_data, measurement_data, time_variable, mesh, log_folder = NULL) {
-    shiny::runApp(priors_shiny(spatial_data = spatial_data, measurement_data = measurement_data, time_variable = time_variable, mesh = mesh, log_folder = log_folder))
+model_builder <- function(spatial_data, measurement_data, time_variable, mesh, log_folder = NULL) {
+    shiny::runApp(model_builder_shiny(spatial_data = spatial_data, measurement_data = measurement_data, time_variable = time_variable, mesh = mesh, log_folder = log_folder))
 }